requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
bleach==6.0.0
bokeh==3.1.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorcet==3.0.1
contourpy==1.1.0
Django==4.1.4
djangorestframework==3.14.0
holoviews==1.16.2
huey==2.4.5
idna==3.4
Jinja2==3.1.2
linkify-it-py==2.0.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdit-py-plugins==0.4.0
mdurl==0.1.2
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
panel==1.1.0
param==1.13.0
Pillow==9.5.0
pyct==0.5.0
python-dateutil==2.8.2
pytz==2023.3
pyviz-comms==2.3.1
PyYAML==6.0
redis==4.5.5
regex==2023.6.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.3
webencodings==0.5.1
xyzservices==2023.5.0
